Added Zelda directories and files.
The most important file for a new user is 'params_explanatory.ini'. It is a
text-only parameter file that can be fed to Zelda with the command
./zelda params_explanatory.ini
Read this documented file to learn about the most important parameters in Zelda.
The file can be also used as a template for creating your custom parameter files.

The directory structure of Zelda is important to learn how the code works:

* The 'source' directory contains the main source files in C. Each file
corresponds to a module of Zelda.

* The 'tool' directory contains accessory source files in C with
purely numerical functions or utility functions.

* The 'main' directory contains the main source files, i.e. the executable
files, including zelda.c.

* The 'python' directory contains Python scripts to launch Zelda, including
the Zelda wrapper (zelda.py), the batch script (zelda_script.py) and a
rebinning function.

* The 'include' directory contains the declaration files (.h) for all the
C files in the 'source', 'main' and 'tools' directories.

* The 'scripts' directory contains accessory script files in bash or
other scripting languages. For example, to fetch catalogues from
remote servers (e.g. the Millennium simulation server) or handling
catalogue files.

* The 'test' directory contains executable programs to test the outputs
of Zelda.